make it a bit more puzzlery

+++ src/types/unsignedFinally/unsignedFinally.kts
@@ -1,11 +1,11 @@
 package types.unsignedFinally
 
 var x: UInt = 0u
-println(x--)
+println(x--.toInt())
 println(--x)
 
 // What will it print?
-// a) 4294967295 4294967294
-// b) 0 4294967294
-// c) 0 -2
+// a) -1 4294967294
+// b)  0 4294967294
+// c)  0 -2
 // d) will not compile
